feat(login): add login hook
